chore(agents): lift chat tui to shared directory
