+++ packages/nemo_platform_ext/src/nemo_platform_ext/cli/chat_tui.py
@@ -1,0 +1,448 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Reusable terminal UI for OpenAI-compatible streaming chat commands."""
+
+from __future__ import annotations
+
+import json
+import logging
+import re
+import sys
+from types import TracebackType
+from typing import Any, Callable, Iterator, Protocol
+
+import click
+from nemo_platform._streaming import SSEDecoder
+from rich.align import Align
+from rich.console import Console
+from rich.live import Live
+from rich.markdown import Markdown
+from rich.panel import Panel
+from rich.prompt import Prompt
+
+from nemo_platform_ext.cli.core.help_formatter import _get_terminal_width
+
+
+class StreamingBody(Protocol):
+    """Byte-streaming HTTP response returned by an SDK context manager."""
+
+    def iter_bytes(self) -> Iterator[bytes]: ...
+
+
+class StreamingResponse(Protocol):
+    """Context manager returned by generated SDK streaming wrappers."""
+
+    def __enter__(self) -> StreamingBody: ...
+
+    def __exit__(
+        self,
+        exc_type: type[BaseException] | None,
+        exc: BaseException | None,
+        exc_tb: TracebackType | None,
+    ) -> bool | None: ...
+
+
+SendTurn = Callable[[str], StreamingResponse]
+RecordAssistantMessage = Callable[[str], None]
+
+
+class _StreamingThinkingFilter:
+    """Incrementally strip thinking tags while preserving visible content."""
+
+    def __init__(self) -> None:
+        self._pending = ""
+        self._in_thinking = False
+
+    def feed(self, text: str) -> str:
+        """Return visible text that is safe to print immediately."""
+        if not text:
+            return ""
+
+        data = self._pending + text
+        self._pending = ""
+        visible = ""
+
+        while data:
+            if self._in_thinking:
+                close_index = data.find("</think>")
+                if close_index >= 0:
+                    data = data[close_index + len("</think>") :]
+                    self._in_thinking = False
+                    continue
+
+                pending_length = _partial_tag_suffix_length(data, "</think>")
+                self._pending = data[-pending_length:] if pending_length else ""
+                break
+
+            open_index = data.find("<think>")
+            if open_index >= 0:
+                visible += data[:open_index]
+                data = data[open_index + len("<think>") :]
+                self._in_thinking = True
+                continue
+
+            pending_length = _partial_tag_suffix_length(data, "<think>")
+            visible += data[:-pending_length] if pending_length else data
+            self._pending = data[-pending_length:] if pending_length else ""
+            break
+
+        return visible
+
+    def finish(self) -> str:
+        """Flush any buffered non-thinking text at the end of the stream."""
+        if not self._pending:
+            return ""
+
+        pending = self._pending
+        self._pending = ""
+        if self._in_thinking:
+            logging.debug("Dropping incomplete thinking content at end of streamed chat response")
+            return ""
+
+        return pending
+
+
+USER_PANEL_WIDTH_RATIO = 0.8
+ASSISTANT_PANEL_WIDTH_RATIO = 0.95
+LIVE_REFRESH_RATE = 20
+ANSI_MOVE_UP_CLEAR_LINE = "\033[1A\033[2K"
+THINKING_TAG_PATTERN = r"<think>(.*?)</think>"
+THINKING_UNCLOSED_PATTERN = r"<think>(.*?)$"
+
+terminal_width = _get_terminal_width()
+console = Console(width=terminal_width)
+
+
+def run_chat_tui(
+    *,
+    send_turn: SendTurn,
+    display_info: dict[str, str],
+    temperature: float | None = None,
+    max_tokens: int | None = None,
+    system_message: str | None = None,
+    initial_message: str | None = None,
+    record_assistant_message: RecordAssistantMessage | None = None,
+) -> None:
+    """Run the shared interactive chat UI against a caller-provided transport."""
+    root_logger = logging.getLogger()
+    initial_level = root_logger.level
+    try:
+        root_logger.setLevel(logging.CRITICAL)
+        _print_welcome_header(display_info, temperature, max_tokens, system_message)
+
+        last_thinking_content = ""
+        thinking_displayed = False
+
+        if initial_message:
+            result = _process_user_message(initial_message, send_turn)
+            if result:
+                assistant_message, last_thinking_content = result
+                if record_assistant_message is not None:
+                    record_assistant_message(assistant_message)
+
+        while True:
+            try:
+                console.print()
+                user_input = Prompt.ask("[bold cyan]You[/bold cyan]", console=console).strip()
+
+                if not user_input:
+                    continue
+
+                if user_input.startswith("/"):
+                    _clear_prompt_line()
+                    new_thinking_state = _handle_special_command(
+                        user_input,
+                        last_thinking_content,
+                        thinking_displayed,
+                    )
+                    if new_thinking_state is not None:
+                        thinking_displayed = new_thinking_state
+                    continue
+
+                _clear_prompt_line()
+                result = _process_user_message(user_input, send_turn)
+                if result:
+                    assistant_message, last_thinking_content = result
+                    thinking_displayed = False
+                    if record_assistant_message is not None:
+                        record_assistant_message(assistant_message)
+
+            except (KeyboardInterrupt, EOFError):
+                _exit_gracefully()
+
+    finally:
+        root_logger.setLevel(initial_level)
+
+
+def collect_stream_response(response: StreamingResponse) -> tuple[str, Any | None]:
+    """Collect content chunks from an OpenAI-compatible streaming response."""
+    raw_message = ""
+    usage = None
+
+    for delta in _iter_stream_deltas(response):
+        if delta.get("usage") is not None:
+            usage = delta["usage"]
+
+        content = _stream_delta_content(delta)
+        if content:
+            raw_message += content
+
+    return raw_message, usage
+
+
+def stream_text_response(response: StreamingResponse) -> None:
+    """Stream plain text output while hiding thinking markup."""
+    thinking_filter = _StreamingThinkingFilter()
+
+    for delta in _iter_stream_deltas(response):
+        content = _stream_delta_content(delta)
+        if not content:
+            continue
+
+        visible_content = thinking_filter.feed(content)
+        if visible_content:
+            click.echo(visible_content, nl=False)
+            sys.stdout.flush()
+
+    trailing_content = thinking_filter.finish()
+    if trailing_content:
+        click.echo(trailing_content, nl=False)
+        sys.stdout.flush()
+    click.echo()
+
+
+def parse_thinking(text: str) -> tuple[str, str]:
+    """Return thinking content and visible content parsed from think tags."""
+    matches = re.findall(THINKING_TAG_PATTERN, text, re.DOTALL)
+    regular_content = re.sub(THINKING_TAG_PATTERN, "", text, flags=re.DOTALL)
+    thinking_parts = list(matches)
+
+    unclosed_match = re.search(THINKING_UNCLOSED_PATTERN, regular_content, re.DOTALL)
+    if unclosed_match:
+        thinking_parts.append(unclosed_match.group(1))
+        regular_content = re.sub(THINKING_UNCLOSED_PATTERN, "", regular_content, flags=re.DOTALL)
+
+    if thinking_parts:
+        return "\n\n".join(thinking_parts), regular_content.strip()
+
+    return "", text
+
+
+def _partial_tag_suffix_length(text: str, tag: str) -> int:
+    max_length = min(len(text), len(tag) - 1)
+    for length in range(max_length, 0, -1):
+        if tag.startswith(text[-length:]):
+            return length
+    return 0
+
+
+def _iter_stream_deltas(response: StreamingResponse) -> Iterator[dict[str, Any]]:
+    """Yield parsed OpenAI-compatible streaming delta payloads."""
+    with response as stream:
+        for event in SSEDecoder().iter_bytes(stream.iter_bytes()):
+            if event.event == "error":
+                raise click.ClickException(_format_streaming_error(stream, event.data))
+            if event.event is not None:
+                continue
+            if not event.data:
+                continue
+            if event.data.startswith("[DONE]"):
+                break
+
+            try:
+                yield json.loads(event.data)
+            except json.JSONDecodeError:
+                logging.debug("Failed to parse JSON stream event: %s", event.data)
+
+
+def _format_streaming_error(response: StreamingBody, data: str) -> str:
+    if data:
+        return f"Streaming chat request failed: {data}"
+
+    status_code = getattr(response, "status_code", None)
+    if isinstance(status_code, int):
+        return f"Streaming chat request failed (HTTP {status_code})"
+    return "Streaming chat request failed"
+
+
+def _stream_delta_content(delta: dict[str, Any]) -> str:
+    choices = delta.get("choices", [])
+    if not choices or not isinstance(choices[0], dict):
+        return ""
+
+    chunk_delta = choices[0].get("delta", {})
+    if not isinstance(chunk_delta, dict):
+        return ""
+
+    content = chunk_delta.get("content")
+    return content if isinstance(content, str) else ""
+
+
+def _process_user_message(user_input: str, send_turn: SendTurn) -> tuple[str, str] | None:
+    panel_width = int(terminal_width * USER_PANEL_WIDTH_RATIO)
+    user_panel = Panel(
+        user_input,
+        title="[bold cyan]You[/bold cyan]",
+        title_align="right",
+        border_style="cyan",
+        padding=(0, 1),
+        width=panel_width,
+    )
+    console.print(Align.right(user_panel))
+
+    response = send_turn(user_input)
+    assistant_message, thinking_content = _stream_response(response)
+
+    if assistant_message:
+        return assistant_message, thinking_content
+
+    logging.warning("Received empty response from API")
+    console.print(Panel.fit("⚠ Received empty response from model", border_style="yellow", padding=(0, 1)))
+    return None
+
+
+def _clear_prompt_line() -> None:
+    console.file.write(ANSI_MOVE_UP_CLEAR_LINE)
+    console.file.flush()
+
+
+def _exit_gracefully() -> None:
+    console.print("\n")
+    console.print(Panel.fit("[bold]Chat session ended[/bold]", border_style="dim", padding=(0, 2)))
+    raise click.exceptions.Exit(0)
+
+
+def _handle_special_command(command: str, last_thinking: str, thinking_displayed: bool) -> bool | None:
+    if command in {"/thinking", "/t"}:
+        if not last_thinking:
+            console.print(Panel.fit("No reasoning content in the last response", border_style="yellow", padding=(0, 1)))
+            return None
+
+        if thinking_displayed:
+            console.print(Panel.fit("Reasoning already displayed above", border_style="yellow dim", padding=(0, 1)))
+            return True
+
+        thinking_panel = Panel(
+            Markdown(last_thinking),
+            title="[bold yellow]💭 Reasoning[/bold yellow]",
+            title_align="left",
+            border_style="yellow dim",
+            padding=(0, 1),
+        )
+        console.print(thinking_panel)
+        return True
+
+    if command in {"/help", "/h"}:
+        console.print(
+            Panel(
+                "[cyan]/thinking[/cyan] - Show/hide model reasoning from last response\n"
+                "[cyan]/help[/cyan] - Show this help message\n"
+                "[cyan]Ctrl+C[/cyan] - Exit the chat",
+                title="[bold]Available Commands[/bold]",
+                border_style="blue",
+                padding=(0, 1),
+            )
+        )
+        return None
+
+    console.print(
+        Panel.fit(
+            f"Unknown command: {command}. Type /help for available commands.",
+            border_style="yellow",
+            padding=(0, 1),
+        )
+    )
+    return None
+
+
+def _print_welcome_header(
+    display_info: dict[str, str],
+    temperature: float | None,
+    max_tokens: int | None,
+    system_message: str | None,
+) -> None:
+    config_lines = [f"[cyan]{key}:[/cyan] {value}" for key, value in display_info.items()]
+
+    if temperature is not None:
+        config_lines.append(f"[cyan]Temperature:[/cyan] {temperature}")
+    if max_tokens is not None:
+        config_lines.append(f"[cyan]Max Tokens:[/cyan] {max_tokens}")
+    if system_message:
+        config_lines.append(f"[cyan]System:[/cyan] {system_message}")
+
+    welcome_panel = Panel(
+        "\n".join(config_lines),
+        title="[bold green]🤖 NeMo Platform Chat Session[/bold green]",
+        subtitle="Press Ctrl+C to exit",
+        border_style="green",
+        padding=(1, 2),
+    )
+    console.print(welcome_panel)
+
+
+def _is_inside_thinking_tag(message: str) -> bool:
+    return message.count("<think>") > message.count("</think>")
+
+
+def _create_thinking_preview(message: str) -> str:
+    thinking_match = re.search(THINKING_UNCLOSED_PATTERN, message, re.DOTALL)
+    current_thinking = thinking_match.group(1).strip() if thinking_match else ""
+
+    if not current_thinking:
+        return "[dim]💭 Thinking...[/dim]"
+
+    lines = current_thinking.split("\n")
+    preview = "\n".join(lines[-3:]) if len(lines) > 1 else current_thinking
+    if len(preview) > 150:
+        preview = "..." + preview[-150:]
+    return f"[dim italic]{preview}[/dim italic]"
+
+
+def _extract_display_text(message: str, inside_thinking: bool) -> str:
+    display_text = re.sub(THINKING_TAG_PATTERN, "", message, flags=re.DOTALL)
+    if inside_thinking:
+        display_text = re.sub(THINKING_UNCLOSED_PATTERN, "", display_text, flags=re.DOTALL)
+    return display_text.strip()
+
+
+def _stream_response(response: StreamingResponse) -> tuple[str, str]:
+    full_message = ""
+    panel_width = int(terminal_width * ASSISTANT_PANEL_WIDTH_RATIO)
+
+    with Live("", console=console, refresh_per_second=LIVE_REFRESH_RATE) as live:
+        for delta in _iter_stream_deltas(response):
+            content = _stream_delta_content(delta)
+            if not content:
+                continue
+
+            full_message += content
+            inside_thinking = _is_inside_thinking_tag(full_message)
+            display_text = _extract_display_text(full_message, inside_thinking)
+
+            if inside_thinking and not display_text:
+                thinking_indicator = Panel(
+                    _create_thinking_preview(full_message),
+                    title="[bold yellow dim]💭 Reasoning[/bold yellow dim]",
+                    title_align="left",
+                    border_style="yellow dim",
+                    padding=(0, 1),
+                    width=panel_width,
+                )
+                live.update(Align.left(thinking_indicator))
+            else:
+                response_panel = Panel(
+                    Markdown(display_text) if display_text else "",
+                    title="[bold magenta]Assistant[/bold magenta]",
+                    title_align="left",
+                    border_style="magenta",
+                    padding=(0, 1),
+                    width=panel_width,
+                )
+                live.update(Align.left(response_panel))
+
+    thinking_content, regular_content = parse_thinking(full_message)
+    if thinking_content:
+        console.print("[dim]💭 Reasoning available! Type [dim cyan]/thinking[/] to view it[/]", markup=True)
+
+    return regular_content or full_message, thinking_content
